파일 및 폴더구조 fsd에 맞게 변경
