+++ src/app/Router.jsx
@@ -1,17 +1,20 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Items from "../pages/Items";
-import Home from "../pages/home/Home";
-import AddItem from "../pages/addItem/AddItem";
+import Items from "../pages/Items/index";
+import Home from "../pages/home/index";
+import AddItem from "../pages/addItem/index";
+import Layout from "../shared/ui/layout/Layout";
 
 const Router = () => {
   return (
     <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/items" element={<Items />} />
-        <Route path="/addItem" element={<AddItem />} />
-      </Routes>
+      <Layout>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/items" element={<Items />} />
+          <Route path="/addItem" element={<AddItem />} />
+        </Routes>
+      </Layout>
     </BrowserRouter>
   );
 };
