+++ src/features/allItem/AllItems.jsx
@@ -1,19 +1,20 @@
 import React, { useState } from "react";
 import AllItem from "./AllItem";
 
-import selectButton from "../../../shared/assets/selectButton.svg";
-import glass from "../../../shared/assets/glass.svg";
+import selectButton from "../../shared/ui/assets/selectButton.svg";
+import glass from "../../shared/ui/assets/glass.svg";
 
-import SearchBar from "./SearchBar";
-import SortSelect from "./SortSelect";
-import SortDropDownList from "./SortDropDownList";
-import AddItemButton from "./AddItemButton";
-import { fetchItems } from "../service/fetchItems";
+import SearchBar from "../search/SearchBar";
+import SortSelect from "../sorting/SortSelect";
+import SortDropDownList from "../sorting/SortDropDownList";
+import AddItemButton from "../addItem/AddItemButton";
+import { fetchItems } from "../../entities/items/fetchItems";
 import { useQuery } from "@tanstack/react-query";
-import ReactPaginate from "react-paginate";
-import useResponsiveDisplayCount from "../hooks/useResponsiveDisplayCount";
+
+import useResponsiveDisplayCount from "../../entities/items/hooks/useResponsiveDisplayCount";
+import Pagination from "../pagination/Pagination";
 const AllItems = () => {
-  const { bestDisplayCount, allDisplayCount } = useResponsiveDisplayCount(); // 훅 사용
+  const { allDisplayCount } = useResponsiveDisplayCount(); // 훅 사용
   const [isOpen, setIsOpen] = useState(false); // 드롭다운 열림/닫힘 상태
   const [currentPage, setCurrentPage] = useState(1);
 
@@ -32,8 +33,6 @@ const AllItems = () => {
     queryKey: ["posts", currentPage, allDisplayCount, orderBy], // 캐시 키는 페이지마다 다르게 설정
     queryFn: () => fetchItems(currentPage, allDisplayCount, orderBy), // 데이터를 가져오는 함수
     keepPreviousData: true, //
-    refetchOnWindowFocus: false,
-    retry: false,
   });
 
   const handlePageClick = (event) => {
@@ -96,27 +95,11 @@ const AllItems = () => {
           <AllItem product={product} key={product.id} />
         ))}
       </div>
-      <ReactPaginate
-        previousLabel={"<"}
-        nextLabel={">"}
-        breakLabel={"..."}
-        pageCount={totalPages} // 총 페이지 수를 설정
-        marginPagesDisplayed={0} // 양쪽에 표시할 페이지 수
-        pageRangeDisplayed={5} // 현재 페이지 기준으로 표시할 페이지 수
-        onPageChange={handlePageClick} // 페이지가 변경될 때 호출되는 함수
-        forcePage={currentPage - 1} // 현재 페이지 상태를 강제로 유지
-        containerClassName={"flex justify-center items-center space-x-2 mt-2"} // 페이지네이션 컨테이너를 가로로 정렬
-        pageClassName={"page-item"} // 페이지 항목 클래스
-        pageLinkClassName={"page-link px-3 py-1 border rounded"} // 페이지 링크 클래스 (버튼 스타일 추가)
-        previousClassName={"page-item"}
-        previousLinkClassName={"page-link px-3 py-1 border rounded"}
-        nextClassName={"page-item"}
-        nextLinkClassName={"page-link px-3 py-1 border rounded"}
-        breakClassName={"page-item"}
-        breakLinkClassName={"page-link px-3 py-1 border rounded"}
-        activeClassName={"active bg-blue-500 text-white"} // 활성화된 페이지 클래스
-      />{" "}
-      {/* 페이지를 바꿀 때 */}
+      <Pagination
+        totalPages={totalPages}
+        currentPage={currentPage}
+        handlePageClick={handlePageClick}
+      />
     </div>
   );
 };
