+++ src/features/bestItem/BestItems.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { fetchItems } from "../service/fetchItems";
+import { fetchItems } from "../../entities/items/fetchItems";
 import BestItem from "./BestItem";
 import { useQuery } from "@tanstack/react-query";
-import useResponsiveDisplayCount from "../hooks/useResponsiveDisplayCount";
+import useResponsiveDisplayCount from "../../entities/items/hooks/useResponsiveDisplayCount";
 
 const BestItems = () => {
   const { bestDisplayCount } = useResponsiveDisplayCount(); // 훅 사용
