--- src/pages/home/Home.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Nav from "../../shared/ui/nav/Nav";
-import { Link } from "react-router-dom";
-
-const Home = () => {
-  return (
-    <div>
-      <Nav />
-      중고마켓 클릭시 items 페이지로 이동
-    </div>
-  );
-};
-
-export default Home;
