+++ src/pages/home/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Home = () => {
+  return <div>중고마켓 클릭시 items 페이지로 이동</div>;
+};
+
+export default Home;
