+++ src/shared/ui/header/LeftNav.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import logo from "../../assets/logo.svg";
-import logoText from "../../assets/logoText.svg";
+import logo from "../assets/logo.svg";
+import logoText from "../assets/logoText.svg";
 import { Link } from "react-router-dom";
 
 const LeftNav = () => {
