+++ src/shared/ui/layout/Layout.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import Nav from "../header/Nav";
+
+const Layout = ({ children }) => {
+  return (
+    <>
+      <header>
+        <Nav />
+      </header>
+      <main>{children}</main>
+    </>
+  );
+};
+
+export default Layout;
